Rethrow thrown errors in checkThrow
When checkThrow is called with a property that fails due to a thrown
Error, instead of throwing a new Error whose message includes the stack
trace (which leads to noisy reporting from test frameworks that print
the message), just update the message on the thrown Error and reuse it.
This way, the outer framework has the choice of including the stack
trace or not, and any additional data attached to the Error aren't lost.