Changed example boolean arbitraries to jsc.bool
Writing some tests myself I was blocked on an error coming back saying `TypeError: jsc.bool is not a function`. After some trial and error I found out that the expected syntax is to use `jsc.bool` instead of calling bool as if it is a function as in `jsc.bool()`.